Extract Hurricane from PressureSystem to a separate subclass
[#163254310]
